Prevent crash in case of multiple entries in glossary. Changed description of glossary-collection.
When there were more than one entries for a glossterm, a warning was emitted and than the transformation failed with a runtime error.
This was changed so that only the first of many entries will be referenced.
Testcase glossary.001 was extended and will cover two entries for one term Pepper.

The description for glossary-collection was changed.
It suggested the necessity for a lot of empty glossentries which must match external entries by xml:id.
Seems to be pretty easy, see test glossary.002.

- just create an empty glossary with role='auto'. That's all
- we dont need @xml:id in glossentries