Make sure bibliography entries have unique IDs
Fix #576

The code that generates unique IDs can only do so within the context of a single
document. If you include entries from multiple documents, collisions can occur.
It turns out that the solution is not to attempt to add IDs during
normalization, but to add IDs during the normal formatting process. This is how
glossary entries work.