check if table width exceeds nominal page width more than treshold
See #609. Switch to relative width occurs only if sum of column widths
exceeds `$nominal-page-width` by more than a `$treshold`, which is
`1px`. This helps to avoid false positive test because of rounding
errors.

Analysis shows a potential issue with `f:absolute-length` in module
`units.xsl`: although the value is rounded without fraction digits,
the function signature gives it the type `xs:double`. I think function
signature should be changed to

```
<xsl:function name="f:absolute-length" as="xs:nonNegativeInteger">
```

If table width really exceeds `nominal-page-width`, the error message
is optimized and informs about the amount of exceeding as well as the
path of the table in question.