Fix broken citation links when entries don’t have an xml:id
The culprit here was that the links were being calculated relative to
their location in the external bibliography. By the time we’re
processing a citation, they’re in the current document.

There’s a small change in markup here, so technically backwards
incompatible, but I’m going to risk it.

Previously, citation produced:

    <a href="#R_bi1_biblioentry1" class="xref xref-biblioentry">
      <span class="label">
        <span class="abbrev">AhoSethiUllman96</span>
      </span>
    </a>

Which I don’t much like anyway; now it produces:

  <span class="citation">
    <a href="#R_bi1_biblioentry1" class="xref xref-biblioentry">AhoSethiUllman96</a>
  </span>

It’s possible that the label/abbrev classes were used for styling, but
styling the citation class should be a lot easier.