Support language localization with Saxon PE
Fix #647

Attempting to generate day or month names in a language other than English
doesn’t work with Saxon HE. But it does work with Saxon PE which was not
supported by the stylesheets.

Note that HE doesn’t fail, it outputs [language: en] to alert you that it
ignored the language. For some reason, I decided to work around this by avoiding
the language localization when using Saxon HE. I’m not sure that was the right
thing to do, but it’s been that way for a long time and it doesn’t seem worth
breaking backwards compatibility over.

Note that it is possible to add localization support to Saxon HE. If you did
that, you’d want a switch to disable the default behavior here. I’m not adding
that until someone asks because I think it’s pretty unlikely.