c# This is a combination of 2 commits.
Backwards incompatible: width/depth percentage values changed

When the width or depth of a media object is specified as a percentage, it is
supposed to be a percentage of the available space. All versions of the
stylesheets before 2.7.0 incorrectly calculate percentages against the intrinsic
size of the +image.

Fixing this is backwards incompatible. In an effort to reduce the impact,
a new parameter $mediaobject-percentage-of-intrinsic-size has been added
which preserves the original, incorrect calculations.