Compute nominal width in CALS tables
This PR properly fixes #581. The stylesheets already know the nominal width of
each column in a CALS table. That information is now passed to the mediaobject
handling code so that the width can be computed correctly.

The new tunnel parameter $inherited-nominal-width could potentially be used in
other contexts to adjust the width of media objects. Perhaps, for example, if we
knew that a multicolumn layout was being used?